sethackeny cookie as a cypress command, refactor set up on search rename unit to pages (gonna break out pages into smaller tests- maybe later)
