+++ cypress/integration/pages/details.spec.js
@@ -138,9 +138,7 @@ describe('Details Page', () => {
         .should('contain', 'Read more')
         .and('not.contain', 'Read less');
     });
-  });
 
-  describe('Read More', () => {
     it('Does not display read more/less button if note is less than 128 characters', () => {
       cy.get('.activity > table > tbody > tr:nth-child(4)')
         .first()
@@ -169,77 +167,6 @@ describe('Details Page', () => {
 
       cy.get('.govuk-details > .govuk-details__text');
       'contain', 'JIGSAW';
-    });
-  });
-
-  describe('Back to search button', () => {
-    it('Back to search takes you back to search after viewing record', () => {
-      setHackneyCookie(true);
-      cy.visit(
-        'http://localhost:3001/search?firstName=wednesday&lastName=adams'
-      );
-      cy.get('.govuk-button')
-        .first()
-        .click();
-      cy.get('.govuk-back-link')
-        .scrollIntoView()
-        .should('contain', 'Back to search')
-        .click();
-
-      cy.get('body').should('contain', 'Customers with matching details');
-    });
-
-    it('Back to search takes you back to search after connecting records', () => {
-      setHackneyCookie(true);
-      cy.visit(
-        'http://localhost:3001/search?firstName=wednesday&lastName=adams'
-      );
-
-      cy.get('.groupedTable')
-        .first()
-        .scrollIntoView()
-        .find('tr')
-        .then(result => {
-          result.each((_, otherThing) => {
-            otherThing.click();
-            console.log(otherThing);
-          });
-        });
-
-      cy.get('.govuk-button')
-        .last()
-        .scrollIntoView()
-        .click({ force: true });
-
-      cy.get('.govuk-back-link')
-        .scrollIntoView()
-        .should('contain', 'Back to search')
-        .click();
-
-      cy.get('body').should('contain', 'Customers with matching details');
-    });
-
-    it('Back to search takes you back to search after viewing record and clicking on more details in quick access', () => {
-      setHackneyCookie(true);
-      cy.visit(
-        'http://localhost:3001/search?firstName=wednesday&lastName=adams'
-      );
-      cy.get('.govuk-button')
-        .first()
-        .click();
-
-      cy.get('.quick-access__item__links')
-        .last()
-        .click();
-
-      cy.get('.close').click();
-
-      cy.get('.govuk-back-link')
-        .scrollIntoView()
-        .should('contain', 'Back to search')
-        .click();
-
-      cy.get('body').should('contain', 'Customers with matching details');
     });
   });
 });
