+++ cypress/support/commands.js
@@ -23,3 +23,18 @@
 //
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+import jwt from 'jsonwebtoken';
+
+const setHackneyCookie =  isValidGroup => {
+  const group = isValidGroup
+    ? 'housingneeds-singleview-beta'
+    : 'some-other-group';
+  const token = jwt.sign({ groups: [group] }, 'a-secure-signature');
+   cy.setCookie('hackneyToken', token, {
+    url: 'http://localhost:3001',
+    domain: 'localhost'
+  });
+};
+
+Cypress.Commands.add('setHackneyCookie', setHackneyCookie);
